+++ NEWS.md
@@ -18,6 +18,7 @@
 * Fix `collection` response JSON `extent` property to be OAPIF-compliant (#54)
 * Fix landing page document to include links to resources `service-desc` and `conformance`
 * Fix links in JSON responses to not include JSON format (required for QGIS)
+* Fix function feature requests to respect the `offset` parameter (#65)
 
 ## Version 1.2
 *Released: 2020 Dec 1*